Enable Bluetooth controller logs on need.
This change will send a vendor specific command to enable/disable
controller logs on need. Once the command is received to the SOC,
It would start sending cotroller's print strings and LMP RX and TX logs
over the UART. There is a property to be set to enable this logging.

Change-Id: I967723ab89a9e0ab49ba10493f3e4a363c32e5af