Bluetooth: add bluetooth and wifi status sync change
In QCA6234, there is high power consumption when BT alone
is turned on because BT requires WLAN module to be out of
reset and WLAN module provides clock to BT. After BT turns
on WLAN's HSIC interface goes into reset mode, which consume
a lot of power. So adding sync between BT and WLAN module to
put hsic into auto suspend mode when BT alone is turned on.

Change-Id: I80fa8c784e95c8fc31ba75554676a72e4ae9fac4